cleanup

+++ plugins/k8s/internal/eventstream/rmq.go
@@ -221,8 +221,8 @@ func (es *EventStream) StartDeleteConsumer(ctx context.Context) error {
 		return fmt.Errorf("rabbitmq connection is closed")
 	}
 
-	log.Debug().Msg("creating daemon_delete_request queue")
 	queueName := "daemon_delete_request"
+	log.Debug().Msgf("creating %v queue for processing checkpoint delete requests", queueName)
 	consumer, err := rabbitmq.NewConsumer(
 		conn,
 		queueName,
@@ -246,16 +246,21 @@ func (es *EventStream) DeleteHandler(ctx context.Context) rabbitmq.Handler {
 			return rabbitmq.Ack
 		}
 
+    if deleteReq.CheckpointPath == "" {
+      log.Error().Msg("request has empty checkpoint path")
+      return rabbitmq.Ack
+    }
+
 		daemonReq := &daemon.DeleteCheckpointReq{
 			Path: &deleteReq.CheckpointPath,
 		}
 
-		log.Debug().Any("deleteReq", deleteReq).Msg("Got request from queue, sending to daemon")
 		_, err := es.cedana.DeleteCheckpoint(ctx, daemonReq)
 		if err != nil {
 			log.Error().Err(err).Msg("could not delete checkpoint")
 			return rabbitmq.NackRequeue
 		}
+		log.Debug().Any("path", deleteReq.CheckpointPath).Msg("processed request from delete queue")
 		return rabbitmq.Ack
 	}
 }
